19/6 Update
- Added FontManager
- Added Menu Navigation
- Added more UI Elements

+++ ArcanaBall/src/Events/GameStateEvent.h
@@ -4,11 +4,11 @@
 struct GameStateEvent : BaseEvent {
 	enum class Type { StartGame, EndGame, ContinueGame, MainMenu, GameOver };
 
-	GameStateEvent(Type type, float payload = 0) :
+	GameStateEvent(Type type, int payload = 0) :
 		type(type),
 		payload(payload)
 	{}
 
 	Type type;
-	float payload{};
+	int payload{};
 };
